requirements.txt
alembic==1.7.7
anyio==3.5.0
appdirs==1.4.4
arabic-reshaper==2.1.3
asgiref==3.5.0
asn1crypto==1.5.1
asttokens==2.0.5
autopep8==1.6.0
backcall==0.2.0
bcrypt==3.2.0
black==22.1.0
branca==0.4.2
Brotli==1.0.9
certifi==2021.5.30
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
cryptography==36.0.2
cssselect2==0.6.0
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.1
dj-database-url==0.5.0
Django==4.0.3
django-crispy-forms==1.14.0
django-import-export==2.8.0
dnspython==2.2.1
ecdsa==0.17.0
email-validator==1.1.3
et-xmlfile==1.1.0
executing==0.8.3
fastapi==0.75.0
filelock==3.0.12
folium==0.12.1.post1
future==0.18.2
greenlet==1.1.2
h11==0.13.0
heroku==0.1.4
html5lib==1.1
httptools==0.2.0
idna==3.3
import-export==0.2.67.dev6
ipython==8.4.0
ipython-genutils==0.2.0
ipython-sql==0.4.1
itsdangerous==2.1.1
jedi==0.18.1
Jinja2==3.0.3
lxml==4.8.0
Mako==1.2.0
MarkupPy==1.14
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mccabe==0.6.1
mypy-extensions==0.4.3
mysqlclient==2.1.0
numpy==1.22.3
odfpy==1.4.1
openpyxl==3.0.10
orjson==3.6.7
oscrypto==1.3.0
pandas==1.4.3
parso==0.8.3
passlib==1.7.4
pathspec==0.9.0
pbr==5.6.0
pickleshare==0.7.5
Pillow==8.3.2
platformdirs==2.5.1
prettytable==0.7.2
prompt-toolkit==3.0.30
psycopg2==2.9.1
pure-eval==0.2.2
pyasn1==0.4.8
pycodestyle==2.8.0
pycparser==2.20
pydantic==1.9.0
pydyf==0.0.3
Pygments==2.12.0
pyHanko==0.13.1
pyhanko-certvalidator==0.19.5
PyMySQL==1.0.2
PyPDF2==1.26.0
PyPDF3==1.0.6
pyphen==0.11.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==0.19.2
python-jose==3.3.0
python-multipart==0.0.5
pytz==2021.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==5.4.1
qrcode==7.3.1
reportlab==3.6.9
requests==2.27.1
rsa==4.8
six==1.15.0
sniffio==1.2.0
SQLAlchemy==1.4.32
sqlparse==0.4.2
stack-data==0.3.0
starlette==0.17.1
svglib==1.3.0
tablib==3.0.0
tinycss2==1.1.0
toml==0.10.2
tomli==2.0.1
tqdm==4.64.0
traitlets==5.3.0
typing_extensions==4.1.1
tzdata==2021.5
tzlocal==4.2
ujson==4.3.0
uritools==4.0.0
urllib3==1.26.6
uvicorn==0.15.0
virtualenv-clone==0.5.4
vrmapi==0.1.3
watchgod==0.8
wcwidth==0.2.5
webencodings==0.5.1
websockets==10.2
whitenoise==5.3.0
wrapt==1.12.1
xhtml2pdf==0.2.7
xlrd==2.0.1
xlwt==1.3.0
zopfli==0.1.8
